+++ balance.go
@@ -9,19 +9,19 @@ type Balance struct {
 	Currency string `json:"currency"`
 
 	// Balance amount.
-	Cash float32 `json:"cash"`
+	Cash float64 `json:"cash"`
 
 	// Market value of all securities in the account in a given currency.
-	MarketValue float32 `json:"marketValue"`
+	MarketValue float64 `json:"marketValue"`
 
 	// Equity as a difference between cash and marketValue properties.
-	TotalEquity float32 `json:"totalEquity"`
+	TotalEquity float64 `json:"totalEquity"`
 
 	// Buying power for that particular currency side of the account.
-	BuyingPower float32 `json:"buyingPower"`
+	BuyingPower float64 `json:"buyingPower"`
 
 	// Maintenance excess for that particular side of the account.
-	MaintenanceExcess float32 `json:"maintenanceExcess"`
+	MaintenanceExcess float64 `json:"maintenanceExcess"`
 
 	// Whether real-time data was used to calculate the above values.
 	IsRealTime bool `json:"isRealTime"`
